pyproject.toml
[project]
name = "promoservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.12",
    "djangorestframework>=3.17.0",
    "drf-yasg>=1.21.15",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.7",
]
